requirements.txt
streamlit
pandas
plotly
gTTS
google-generativeai